+++ storage/src/testFixtures/java/tech/pegasys/teku/storage/storageSystem/StorageSystem.java
@@ -57,7 +57,8 @@ public class StorageSystem implements AutoCloseable {
       final RecentChainData recentChainData,
       final CombinedChainDataClient combinedChainDataClient,
       final RestartedStorageSupplier restartedSupplier,
-      final ChainBuilder chainBuilder) {
+      final ChainBuilder chainBuilder,
+      final Spec spec) {
     this.metricsSystem = metricsSystem;
     this.chainStorage = chainStorage;
     this.recentChainData = recentChainData;
@@ -68,7 +69,7 @@ public class StorageSystem implements AutoCloseable {
     this.restartedSupplier = restartedSupplier;
 
     this.chainBuilder = chainBuilder;
-    chainUpdater = new ChainUpdater(this.recentChainData, this.chainBuilder);
+    chainUpdater = new ChainUpdater(this.recentChainData, this.chainBuilder, spec);
   }
 
   static StorageSystem create(
@@ -114,7 +115,8 @@ public class StorageSystem implements AutoCloseable {
         recentChainData,
         combinedChainDataClient,
         restartedSupplier,
-        chainBuilder);
+        chainBuilder,
+        spec);
   }
 
   public StubMetricsSystem getMetricsSystem() {